// src/models/data.d.ts
export declare const Day1: mongoose.Model<{
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
}, {}, {}, {
    id: string;
}, mongoose.Document<unknown, {}, {
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
}, {
    id: string;
}, mongoose.DefaultSchemaOptions> & Omit<{
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
} & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}, "id"> & {
    id: string;
}, mongoose.Schema<any, mongoose.Model<any, any, any, any, any, any, any>, {}, {}, {}, {}, mongoose.DefaultSchemaOptions, {
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
}, mongoose.Document<unknown, {}, {
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
}, {
    id: string;
}, mongoose.DefaultSchemaOptions> & Omit<{
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
} & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}, "id"> & {
    id: string;
}, unknown, {
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
} & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}>, {
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
} & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}>;
export declare const Day2: mongoose.Model<{
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
}, {}, {}, {
    id: string;
}, mongoose.Document<unknown, {}, {
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
}, {
    id: string;
}, mongoose.DefaultSchemaOptions> & Omit<{
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
} & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}, "id"> & {
    id: string;
}, mongoose.Schema<any, mongoose.Model<any, any, any, any, any, any, any>, {}, {}, {}, {}, mongoose.DefaultSchemaOptions, {
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
}, mongoose.Document<unknown, {}, {
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
}, {
    id: string;
}, mongoose.DefaultSchemaOptions> & Omit<{
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
} & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}, "id"> & {
    id: string;
}, unknown, {
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
} & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}>, {
    createdAt: NativeDate;
    isApprovalMode: boolean;
    nik?: string;
    nama?: string;
    jabatan?: string;
    type?: string;
    tanggal?: string;
    tandaTangan?: string;
    waktuMulai?: string;
    waktuSelesai?: string;
    money?: number;
    jam16?: boolean;
    jam12?: boolean;
} & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}>;