// src/models/listUser.d.ts
import mongoose from 'mongoose';
export declare const ListUser: mongoose.Model<{
    updatedAt: NativeDate;
    nik: string;
}, {}, {}, {
    id: string;
}, mongoose.Document<unknown, {}, {
    updatedAt: NativeDate;
    nik: string;
}, {
    id: string;
}, mongoose.DefaultSchemaOptions> & Omit<{
    updatedAt: NativeDate;
    nik: string;
} & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}, "id"> & {
    id: string;
}, mongoose.Schema<any, mongoose.Model<any, any, any, any, any, any, any>, {}, {}, {}, {}, mongoose.DefaultSchemaOptions, {
    updatedAt: NativeDate;
    nik: string;
}, mongoose.Document<unknown, {}, {
    updatedAt: NativeDate;
    nik: string;
}, {
    id: string;
}, mongoose.DefaultSchemaOptions> & Omit<{
    updatedAt: NativeDate;
    nik: string;
} & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}, "id"> & {
    id: string;
}, unknown, {
    updatedAt: NativeDate;
    nik: string;
} & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}>, {
    updatedAt: NativeDate;
    nik: string;
} & {
    _id: mongoose.Types.ObjectId;
} & {
    __v: number;
}>;
//# sourceMappingURL=listUser.d.ts.map